Fielder target selection
Added the possibility to chose the target to pass the ball for the pitcher and the fielders.
Added the selection manager.
Added a new ui element for highlight the players.
Refactored some functions.
Cleaned up code.
Added the tagout selection and the simple version of the target selection for tagout.